Add error handling
This commit revives and adds error handling to the `opml-decode'
function. Errors are signaled if we cannot locate ompl2org.py /or/ if the
process of decoding the opml fails. On success, we return the end-location of
the buffer because that's what format-alist wants according to [1].

Also added autoloads, I /think/ we need these.

[1]: https://www.gnu.org/software/emacs/manual/html_node/elisp/Format-Conversion-Round_002dTrip.html
(I could not find this in my local info with C-h i, dunno why.)

Fixes #5